Restructure project modules
